[신윤식] -20250809 생일인데 아침부터 이 작업을 하고 있다. 좋아요 설정과  테이블 수정에 의한 오류가 발생하였고, comment view 타임리프에서 사용자의 ID가 숫자로 나오는 것을 발견하여  DTO에 UserName 을 추가 하여 다시 리팩토링 진행.... 기다려라 LIst 재료 절차 넣기야...

+++ src/main/java/com/andy/recipe/like/domain/Like.java
@@ -5,7 +5,7 @@ import java.sql.Timestamp;
 public class Like {
 	private long id;
 	private long userId;
-	private int postId;
+	private long postId;
 	private Timestamp createdAt;
 
 	public long getId() {
@@ -24,11 +24,11 @@ public class Like {
 		this.userId = userId;
 	}
 
-	public int getPostId() {
+	public long getPostId() {
 		return postId;
 	}
 
-	public void setPostId(int postId) {
+	public void setPostId(long postId) {
 		this.postId = postId;
 	}
